refactor(transport): extract ClientTransportBase to decouple WS from TCP
Introduce abstract `ClientTransportBase extends TCP_transport` that owns
the UACP HEL/ACK handshake, the negotiated transport state, and the
post-connect connection-break detector. `ClientTCP_transport` is now a
thin TCP-only subclass; the browser `ClientWS_transport` (in
`node-opcua-client-browser`) becomes a sibling rather than a subclass.

This removes the `node:net`/`node:os`/`node:util` imports from the WS
transport's static dependency graph, and lets `_perform_HEL_ACK_transaction`
be `protected` (was `private`) so subclasses no longer need type-cast hacks.

`TransportSettingsOptions` moves to `i_client_transport.ts` (interface
contract, not implementation) with a re-export from `client_tcp_transport.ts`
for source-level backwards compat. The newly extracted
`_install_post_connect_error_handler()` DRYs a regex listener that was
duplicated in both TCP and WS subclasses.

Also: change `message_builder_base.ts` to deep-import `get_clock_tick`
from `node-opcua-utils/dist/get_clock_tick` to bypass the package barrel
that re-exports `check_file_exists` (which `require("fs")`). Lets browser
bundlers omit `fs` from the transport chain entirely.

Behaviour unchanged. Public API of `ClientTCP_transport` preserved.
All existing tests pass (84 in transport, 2 in secure-channel factory seam).
