+++ packages/node-opcua-transport/source/client_tcp_transport.ts
@@ -8,28 +8,27 @@ import { types } from "node:util";
 import chalk from "chalk";
 
 import { assert } from "node-opcua-assert";
-import { BinaryStream } from "node-opcua-binary-stream";
-import { readMessageHeader } from "node-opcua-chunkmanager";
-import { checkDebugFlag, make_debugLog, make_errorLog, make_warningLog } from "node-opcua-debug";
+import { checkDebugFlag, make_debugLog, make_errorLog } from "node-opcua-debug";
 import type { ErrorCallback } from "node-opcua-status-code";
-import { AcknowledgeMessage } from "./AcknowledgeMessage";
-import { HelloMessage } from "./HelloMessage";
-import { TCPErrorMessage } from "./TCPErrorMessage";
-import { getFakeTransport, type ISocketLike, TCP_transport } from "./tcp_transport";
-import { decodeMessage, packTcpMessage, parseEndpointUrl } from "./tools";
-import { doTraceHelloAck } from "./utils";
+
+import { ClientTransportBase } from "./client_transport_base";
+import type { TransportSettingsOptions } from "./i_client_transport";
+import { getFakeTransport, type ISocketLike } from "./tcp_transport";
+import { parseEndpointUrl } from "./tools";
 
 const doDebug = checkDebugFlag(__filename);
 const debugLog = make_debugLog(__filename);
-const warningLog = make_warningLog(__filename);
 const errorLog = make_errorLog(__filename);
 const gHostname = os.hostname();
+
+// Re-exported for source-level backwards compatibility: this type used to live in this file.
+export type { TransportSettingsOptions };
 
 function createClientSocket(endpointUrl: string, timeout: number): ISocketLike {
     // create a socket based on Url
     const ep = parseEndpointUrl(endpointUrl);
-    const port = parseInt(ep.port!, 10);
-    const hostname = ep.hostname!;
+    const port = parseInt(ep.port  || "4840", 10);
+    const hostname = ep.hostname;
 
     let socket: ISocketLike;
     switch (ep.protocol) {
@@ -52,29 +51,6 @@ function createClientSocket(endpointUrl: string, timeout: number): ISocketLike {
             throw new Error(msg);
         }
     }
-}
-export interface ClientTCP_transport {
-    on(eventName: "chunk", eventHandler: (messageChunk: Buffer) => void): this;
-    on(eventName: "close", eventHandler: (err: Error | null) => void): this;
-    on(eventName: "connection_break", eventHandler: (err: Error | null) => void): this;
-    on(eventName: "connect", eventHandler: () => void): this;
-
-    once(eventName: "chunk", eventHandler: (messageChunk: Buffer) => void): this;
-    once(eventName: "close", eventHandler: (err: Error | null) => void): this;
-    once(eventName: "connection_break", eventHandler: (err: Error | null) => void): this;
-    once(eventName: "connect", eventHandler: () => void): this;
-
-    emit(eventName: "chunk", messageChunk: Buffer): boolean;
-    emit(eventName: "close", err?: Error | null): boolean;
-    emit(eventName: "connection_break", err?: Error | null): boolean;
-    emit(eventName: "connect"): boolean;
-}
-
-export interface TransportSettingsOptions {
-    maxChunkCount?: number;
-    maxMessageSize?: number;
-    receiveBufferSize?: number;
-    sendBufferSize?: number;
 }
 
 /**
@@ -112,50 +88,7 @@ export interface TransportSettingsOptions {
  *
  *
  */
-// biome-ignore lint/suspicious/noUnsafeDeclarationMerging: typed EventEmitter events - extending parent's event map in subclass hierarchy
-export class ClientTCP_transport extends TCP_transport {
-    public static defaultMaxChunk = 0; // 0 - no limits
-    public static defaultMaxMessageSize = 0; // 0 - no limits
-    public static defaultReceiveBufferSize = 1024 * 64 * 10;
-    public static defaultSendBufferSize = 1024 * 64 * 10; // 8192 min,
-
-    public endpointUrl: string;
-    public serverUri: string;
-    public numberOfRetry: number;
-    public parameters?: AcknowledgeMessage;
-
-    private _counter: number;
-    private _helloSettings: {
-        maxChunkCount: number;
-        maxMessageSize: number;
-        receiveBufferSize: number;
-        sendBufferSize: number;
-    };
-    constructor(transportSettings?: TransportSettingsOptions) {
-        super();
-        this.endpointUrl = "";
-        this.serverUri = "";
-        this._counter = 0;
-        this.numberOfRetry = 0;
-
-        // initially before HEL/ACK
-        this.maxChunkCount = 1;
-        this.maxMessageSize = 4 * 1024;
-        this.receiveBufferSize = 4 * 1024;
-
-        transportSettings = transportSettings || {};
-        this._helloSettings = {
-            maxChunkCount: transportSettings.maxChunkCount || ClientTCP_transport.defaultMaxChunk,
-            maxMessageSize: transportSettings.maxMessageSize || ClientTCP_transport.defaultMaxMessageSize,
-            receiveBufferSize: transportSettings.receiveBufferSize || ClientTCP_transport.defaultReceiveBufferSize,
-            sendBufferSize: transportSettings.sendBufferSize || ClientTCP_transport.defaultSendBufferSize
-        };
-    }
-
-    public getTransportSettings(): TransportSettingsOptions {
-        return this._helloSettings;
-    }
-
+export class ClientTCP_transport extends ClientTransportBase {
     public dispose(): void {
         /* c8 ignore next */
         doDebug && debugLog(" ClientTCP_transport disposed");
@@ -170,7 +103,8 @@ export class ClientTCP_transport extends TCP_transport {
         doDebug && debugLog(chalk.cyan(`ClientTCP_transport#connect(endpointUrl = ${endpointUrl})`));
         let socket: ISocketLike | null = null;
         try {
-            const _ep = parseEndpointUrl(endpointUrl);
+            // validate the URL upfront so a parse error is reported synchronously
+            parseEndpointUrl(endpointUrl);
             socket = createClientSocket(endpointUrl, this.timeout);
 
             socket.setTimeout(this.timeout >> 1, () => {
@@ -184,34 +118,6 @@ export class ClientTCP_transport extends TCP_transport {
             return;
         }
 
-        /**
-         *
-         */
-        const _on_socket_error_after_connection = (err: Error) => {
-            /* c8 ignore next */
-            doDebug && debugLog(" _on_socket_error_after_connection ClientTCP_transport Socket Error", err.message);
-
-            // EPIPE : EPIPE (Broken pipe): A write on a pipe, socket, or FIFO for which there is no process to read the
-            // data. Commonly encountered at the net and http layers, indicative that the remote side of the stream
-            // being written to has been closed.
-
-            // ECONNRESET (Connection reset by peer): A connection was forcibly closed by a peer. This normally results
-            // from a loss of the connection on the remote socket due to a timeout or reboot. Commonly encountered
-            // via the http and net module
-
-            //  socket termination could happen:
-            //   * when the socket times out (lost of connection, network outage, etc...)
-            //   * or, when the server abruptly disconnects the socket ( in case of invalid communication for instance)
-            if (err.message.match(/ECONNRESET|EPIPE|premature socket termination/)) {
-                /**
-                 * @event connection_break
-                 *
-                 */
-                doDebug && debugLog("connection_break after reconnection", endpointUrl);
-                this.emit("connection_break", err);
-            }
-        };
-
         const _on_socket_connect = () => {
             /* c8 ignore next */
             doDebug && debugLog("entering _on_socket_connect");
@@ -223,13 +129,12 @@ export class ClientTCP_transport extends TCP_transport {
                     if (!this._socket) {
                         return callback(new Error("Abandoned"));
                     }
-                    // install error handler to detect connection break
-                    this._socket.on("error", _on_socket_error_after_connection);
+                    // install the post-connect "connection break" detector inherited from ClientTransportBase
+                    this._install_post_connect_error_handler(endpointUrl);
                     /**
                      * notify the observers that the transport is connected (the socket is connected and the the HEL/ACK
                      * transaction has been done)
                      * @event connect
-                     *
                      */
                     this.emit("connect");
                 } else {
@@ -268,111 +173,5 @@ export class ClientTCP_transport extends TCP_transport {
         this._socket?.once("error", _on_socket_error_for_connect);
         this._socket?.once("end", _on_socket_end_for_connect);
         this._socket?.once("connect", _on_socket_connect);
-    }
-
-    private _handle_ACK_response(messageChunk: Buffer, callback: ErrorCallback) {
-        const _stream = new BinaryStream(messageChunk);
-        const messageHeader = readMessageHeader(_stream);
-        let err: Error | null = null;
-        /* c8 ignore next */
-        if (messageHeader.isFinal !== "F") {
-            err = new Error(" invalid ACK message");
-            return callback(err);
-        }
-
-        let responseClass: typeof AcknowledgeMessage | typeof TCPErrorMessage;
-        let response: AcknowledgeMessage | TCPErrorMessage;
-
-        if (messageHeader.msgType === "ERR") {
-            responseClass = TCPErrorMessage;
-            _stream.rewind();
-            response = decodeMessage(_stream, responseClass) as TCPErrorMessage;
-
-            err = new Error(`ACK: ERR received ${response.statusCode.toString()} : ${response.reason}`);
-            (err as any).statusCode = response.statusCode;
-            // c8 ignore next
-            doTraceHelloAck && warningLog("receiving ERR instead of Ack", response.toString());
-
-            callback(err);
-        } else {
-            responseClass = AcknowledgeMessage;
-            _stream.rewind();
-            response = decodeMessage(_stream, responseClass) as AcknowledgeMessage;
-
-            this.parameters = response;
-            this.setLimits(response);
-
-            // c8 ignore next
-            doTraceHelloAck && warningLog("receiving Ack\n", response.toString());
-
-            callback();
-        }
-    }
-
-    private _send_HELLO_request() {
-        /* c8 ignore next */
-        doDebug && debugLog("entering _send_HELLO_request");
-
-        assert(this._socket);
-        assert(Number.isFinite(this.protocolVersion));
-        assert(this.endpointUrl.length > 0, " expecting a valid endpoint url");
-
-        const { maxChunkCount, maxMessageSize, receiveBufferSize, sendBufferSize } = this._helloSettings;
-
-        // Write a message to the socket as soon as the client is connected,
-        // the server will receive it as message from the client
-        const helloMessage = new HelloMessage({
-            endpointUrl: this.endpointUrl,
-            protocolVersion: this.protocolVersion,
-            maxChunkCount,
-            maxMessageSize,
-            receiveBufferSize,
-            sendBufferSize
-        });
-        // c8 ignore next
-        doTraceHelloAck && warningLog(`sending Hello\n ${helloMessage.toString()} `);
-
-        const messageChunk = packTcpMessage("HEL", helloMessage);
-        this._write_chunk(messageChunk);
-    }
-
-    private _on_ACK_response(externalCallback: ErrorCallback, err: Error | null, data?: Buffer) {
-        /* c8 ignore next */
-        doDebug && debugLog("entering _on_ACK_response");
-
-        assert(typeof externalCallback === "function");
-        assert(this._counter === 0, "Ack response should only be received once !");
-        this._counter += 1;
-
-        if (err || !data) {
-            if (this._socket) {
-                const s = this._socket;
-                this._socket = null;
-                s.destroy();
-            }
-            externalCallback(err || new Error("no data"));
-        } else {
-            this._handle_ACK_response(data, externalCallback);
-        }
-    }
-
-    private _perform_HEL_ACK_transaction(callback: ErrorCallback) {
-        /* c8 ignore next */
-        if (!this._socket) {
-            return callback(new Error("No socket available to perform HEL/ACK transaction"));
-        }
-        assert(this._socket, "expecting a valid socket to send a message");
-        assert(typeof callback === "function");
-        this._counter = 0;
-        /* c8 ignore next */
-        doDebug && debugLog("entering _perform_HEL_ACK_transaction");
-
-        this._install_one_time_message_receiver((err: Error | null, data?: Buffer) => {
-            /* c8 ignore next */
-            doDebug && debugLog("before  _on_ACK_response ", err ? err.message : "");
-
-            this._on_ACK_response(callback, err, data);
-        });
-        this._send_HELLO_request();
     }
 }
